requirements.txt
numpy
pytorch
Hiddenlayer
torchstat
mayavi
